epochStart + epochEnd
this update exposes params for `epochStart` + `epochEnd` ... any events that are BEFORE start or AFTER end will not be imported.

this makes it easier to iterate over files when you don't want to load the entire file but have bounded dates for the import.

times should be provided as a 10-digit unix epoch timestamp (number).